Removed unnecessary abstract classes and removed Abstract keywords if not necessary. Moved LittleEndianDataOutput/InputStream to util package
