+++ src/main/java/net/tammon/sip/packets/utils/LittleEndianDataInputStream.java
@@ -23,7 +23,7 @@
  * SOFTWARE.
  */
 
-package net.tammon.sip.packets.parts;
+package net.tammon.sip.packets.utils;
 
 import java.io.*;
